Coding Journey for Preschooler Volume 1.0 (ScratchJr Installation Guide)
This describes importance of coding for kindergarder and junior school students (3 to 7 years.) .It's designed to develop cognitive abilities, such as perception, logical thinking, and visual processing in children. 
These skills can lead to better critical thinking, problem-solving, and an understanding of mathematics and helps children learn to think creatively, reason systematically, and work collaboratively.