Create InterfaceGuide:[Vol 1.2]  https://designrr.page/?id=484741&token=506021053&h=3730
Interface is a mechanism that allows different software components or systems to communicate and exchange data.There are four main elements of Scratch: the stage, the sprites, the script and the programming palette. These elements can be compared to a play. 1) Stage – similar to the stage in a play. This is where everything will take place.